Add Gemfile to load dependencies while developing
When I ran the tests initially they failed with an error because of a
missing dependency. After installing the dependency I got another error
because the version conflicted with the version that the code in the gem
depends on.

Run bundle install to install gemspec dependencies and use those while
developing and running tests.

Use `bundle exec rake test` to run the test suite using the correct
dependencies.