Update countries gem from 7.0 => 8.0.3
We need to update the countries gem in the Charlie app but it is limited
to 7.0 because of the national holidays gem. So we are updating our
national holidays gem to use the latest version of the countries gem.

[changelog](https://github.com/countries/countries/blob/master/CHANGELOG.md)

Breaking changes:
* Drops support for Ruby 3.1 so have updated the require version to 3.4
* Changes to translation hashes to sometimes use symbol keys instead of
  strings (see commit
[here](https://github.com/countries/countries/pull/884))
All tests in this gem are still passing, however I will make sure to
carefully test the code related to public holidays in the Charlie app before
updating the countries gem there too.

* There is also a potential breaking change in the way the
  `find_subdivision_by_name` method works, but we don't use that in this
gem or the Charlie app.

Finally, this commit also updates the code to consistently use
ISO3166::Country#common_name instead of ISO3166::Country#iso_short_name,
as we were using both across the gem, and common_name returns much more
readable country names.